tried using -O2 instead of -Os
 ... as space is not critical use -O2 instead, BUT this fails
completely! Won't start anymore, so the code is buggy at some
point not allowing for this kind of optimization.

So, onl make some of the time critical functions inlineable,
i.e. put them in the headers and make them static inline.

This is in preparation for the reduction of interrupt latencies
in onewire.c, which block interrupts for quite some time (as
identified and modified by radensb).