sensor.c: let sensor be configured, remove auto-detect
This is in preparation for the implementation of a new control strategy.
Auto-detection of sensors may be neat, but has it's drawbacks
if sensors disappear during operation. Currently this is not
fixed with the change either, but it will allow such a change
more easily.

This also improves the performance of Sensor_DoConversion() as
it will not check for the existance of sensors during the reflow process.