tweak cooling phase with LR_WEIGHTED_AVERAGE
When simulating the PCB temperature the radiation pattern seems
to overheat the PCB even with the lowest weight selected, if
the PCB is placed on stand-offs in the drawer.

But it works super stable having them placed on the drawer bottom
directly! This spreads the heat on the board aswell!
The drawback is the cooling is lagging a lot, even slower
than the heavy thermo couple (which is more exposed to the
air stream caused by the fan).

The solution so far:
- place the PCB directly on the drawer bottom
- with pretty normal PCBs select a weight of about 20%
- remove the cooling ramp from the profiles
  this makes the reflow state machine go to COOLING and fully
  ventilate the drawer.
- tweak the control temperature so that it reflects something
  more likely to be the real PCB temperature (mostly informational
  anyway).
- show the cooling phase on the display

The best solution definitely would be to directly measure the
PCB temperature, but ramp-up with the simulator is about +/-5K
on the setpoint (the PID control still not fixed). The cooling
process is very dependent on the position in the drawer, it's
a lot more air on the left side!