lcd.c : Fix reboot with too long string
The probleme occur when right tc offet is >= 10.00 or <=-10.00
or when version string is more than 21 char as v0.5.2-4-dirty-gxxxxxx.
This cause reboot of the firmware.

The buffer overflow is triggered when trying to print the null char at
the end of the string.

note from snprintf(3) :
snprintf return the return value is the number of characters (excluding the terminating null byte) which would have  been
written  to the final string if enough space had been available.